requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pyaudio==0.2.11
speechrecognition==3.10.0
pynput==1.7.6
tkinter-tooltip==2.0.0
Pillow==10.0.0
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
pytesseract==0.3.10
mss==9.0.1
keyboard==0.13.5
mouse==0.7.1
sounddevice==0.4.6
scipy==1.11.1
librosa==0.10.1
transformers==4.35.0
torch==2.1.0
torchaudio==2.1.0 